requirements.txt
customtkinter
requests
openai-whisper
yt-dlp